queries refactor
