+++ src/queries/all-pools.ts
@@ -1,0 +1,238 @@
+//-----------------------------------------------------------------------------
+// # PRICE HISTORY
+//-----------------------------------------------------------------------------
+
+const PRICE_HISTORY = `
+  usdTVL
+  baseTVL
+  supply
+  absPNL
+  percPNL
+  timestamp
+  aggregationType
+`
+
+const PRICE_HISTORY_LAST = `
+  priceHistory(
+    first: 1,
+    orderBy: timestamp,
+    orderDirection: desc,
+    where: {
+      isLast: true,
+      aggregationType_gte: 0,
+      aggregationType_lte: 0,
+    }) {
+    ${PRICE_HISTORY}
+  }
+`
+
+const PRICE_HISTORY_FULL = (startDate) => `
+  priceHistory(
+    first: $limit, 
+    orderBy: timestamp, orderDirection: asc, 
+    where: { 
+      aggregationType_gte: $minTimeframe,
+      aggregationType_lte: $maxTimeframe,
+      ${startDate !== 0 ? "timestamp_gte: $startDate" : ""}
+    }
+  ) {
+    ${PRICE_HISTORY}
+  }
+`
+
+const PriceHistoryQuery = (startDate) => `
+  query ($address: String!, $minTimeframe: Int!, $maxTimeframe: Int!, $limit: Int!, $startDate: Int!) {
+    traderPool(id: $address) {
+      ${PRICE_HISTORY_FULL(startDate)}
+    }
+  }
+`
+
+//-----------------------------------------------------------------------------
+// # POOL
+//-----------------------------------------------------------------------------
+
+const POOL = `
+  id
+  baseToken
+  name
+  type
+  ticker
+  trader
+  creationTime
+  descriptionURL
+  maxLoss
+  totalTrades
+  totalClosedPositions
+  averageTrades
+  averagePositionTime
+  investorsCount
+  admins
+  privateInvestors {
+    id
+    activePools
+    allPools
+  }
+  ${PRICE_HISTORY_LAST}
+`
+
+const PoolQuery = `
+  query ($address: String!) {
+    traderPool(id: $address) {
+      ${POOL}
+    }
+  }
+`
+
+const OwnedPoolsQuery = `
+  query ($address: String!) {
+    traderPools(orderBy: creationTime, first: 100, where: { trader: $address }) {
+      ${POOL}
+    }
+  }
+`
+const ManagedPoolsQuery = `
+  query ($address: String!) {
+    traderPools(
+      orderBy: creationTime, first: 100, 
+      where: { 
+        admins_contains_nocase: [$address],
+        trader_not: $address
+      }
+    ) {
+      ${POOL}
+    }
+  }
+`
+
+const PoolsQuery = `
+  query ($q: String!) {
+    traderPools(where: { ticker_contains_nocase: $q } first: 100) {
+      ${POOL}
+    }
+  }
+`
+
+const PoolsQueryWithSort = `
+  query ($q: String!, $orderBy: String!, $orderDirection: String!) {
+    traderPools(where: { ticker_contains_nocase: $q } first: 100 orderBy: $orderBy orderDirection: $orderDirection) {
+      ${POOL}
+    }
+  }
+`
+
+const PoolsQueryByType = `
+  query ($q: String!, $type: String!) {
+    traderPools(where: { ticker_contains_nocase: $q, type: $type } first: 100) {
+      ${POOL}
+    }
+  }
+`
+
+const PoolsQueryByTypeWithSort = `
+  query ($q: String!, $type: String!, $orderBy: String!, $orderDirection: String!) {
+    traderPools(where: { ticker_contains_nocase: $q, type: $type } first: 100 orderBy: $orderBy orderDirection: $orderDirection) {
+      ${POOL}
+    }
+  }
+`
+
+// Basic pool positions
+const POSITION_EXCHANGE = `
+  id
+  hash
+  fromToken
+  toToken
+  fromVolume
+  toVolume
+  usdVolume
+  timestamp
+  opening
+`
+
+const POSITION = `
+  id
+  closed
+  positionToken
+  totalUSDOpenVolume
+  totalUSDCloseVolume
+  totalBaseOpenVolume
+  totalBaseCloseVolume
+  totalPositionOpenVolume
+  totalPositionCloseVolume
+  exchanges(orderBy: timestamp, orderDirection: desc) {
+    ${POSITION_EXCHANGE}
+  }
+  traderPool {
+    id
+    trader
+    ticker
+    baseToken
+    descriptionURL
+  }
+`
+
+const BasicPositionsQuery = `
+  query ($offset: Int!, $limit: Int!, $address: String!, $closed: Boolean!) {
+    positions(
+      skip: $offset, first: $limit, 
+      where: { 
+        closed: $closed, 
+        traderPool_: { id: $address }
+      },
+      orderBy: startTimestamp, orderDirection: desc
+    ) {
+      ${POSITION}
+    }
+  }
+`
+
+const PoolPositionLast = `
+query ($poolId: String!, $tokenId: String!) {
+  positions(
+    first: 1, 
+    where: {
+      traderPool_: { id: $poolId }
+      positionToken: $tokenId
+    },
+    orderBy: startTimestamp, orderDirection: desc
+  ) {
+    ${POSITION}
+  }
+}
+`
+
+// Fund fee history
+const FundFeeHistoryQuery = `
+  query($offset: Int!, $limit: Int!, $address: String!) {
+    feeHistories(
+      skip: $offset, first: $limit, 
+      orderBy: day, orderDirection: desc, 
+      where: { traderPool_: { id: $address }}
+    ) {
+      id
+      PNL
+      day
+      fundProfit
+      perfomanceFee
+      traderPool {
+        id
+        baseToken
+      }
+    }
+  }
+`
+
+export {
+  PoolQuery,
+  PoolsQuery,
+  PoolsQueryByType,
+  PriceHistoryQuery,
+  OwnedPoolsQuery,
+  ManagedPoolsQuery,
+  PoolsQueryWithSort,
+  PoolsQueryByTypeWithSort,
+  PoolPositionLast,
+  BasicPositionsQuery,
+  FundFeeHistoryQuery,
+}
