+++ src/queries/investors.ts
@@ -1,0 +1,44 @@
+// Investor positions
+const INVESTOR_POSITION_VEST = `
+  id
+  hash
+  isInvest
+  timestamp
+  volumeBase
+  volumeLP
+  volumeUSD
+`
+
+const InvestorPositionsQuery = `
+  query ($address: String!, $closed: Boolean!, $offset: Int!, $limit: Int!) {
+    investorPoolPositions(skip: $offset, first: $limit, where: {investor: $address, isClosed: $closed}) {
+      id
+      isClosed
+      totalBaseInvestVolume
+      totalBaseDivestVolume
+      totalLPInvestVolume
+      totalLPDivestVolume
+      totalUSDInvestVolume
+      totalUSDDivestVolume
+      pool {
+        id
+        type
+        token
+      }
+      vest(first: 100) {
+        ${INVESTOR_POSITION_VEST}
+      }
+    }
+  }
+`
+
+// Investor proposals
+const InvestorPoolsInvestedForQuery = `
+  query ($address: String!, $poolType: String!) {
+    investors(where: { id: $address }) {
+      activePools(where: { type: $poolType }) { id }
+    }
+  }
+`
+
+export { InvestorPositionsQuery, InvestorPoolsInvestedForQuery }
